src/mlip_autopipec/modules/d_training_engine.py
import subprocess
from pathlib import Path

import numpy as np
from ase.atoms import Atoms
from ase.io import write

from mlip_autopipec.data.database import AseDB
from mlip_autopipec.data.models import TrainingConfig
from mlip_autopipec.utils import baseline_potentials


class TrainingEngine:
    """
    This class encapsulates the logic for training a Machine Learning Interatomic
    Potential (MLIP) model. It handles data loading, preparation for Delta

    Learning, and interfacing with the underlying MLIP framework (e.g., MACE)
    via its command-line interface.
    """

    def __init__(self, config: TrainingConfig, db: AseDB):
        self._config = config
        self._db = db
        self._work_dir = Path.cwd()
        # MACE saves the model in the current working directory based on the 'name'
        self._model_name = "trained_model"
        self._model_path = self._work_dir / f"{self._model_name}.model"


    def execute(self, ids: list[int]) -> Path:
        training_data = self._load_and_prepare_data(ids)
        train_file_path = self._work_dir / "temp_train.xyz"

        try:
            # Save the prepared data to a temporary file for the CLI tool
            write(train_file_path, training_data)

            # Construct the command-line arguments for mace_run_train
            command = [
                "mace_run_train",
                f"--name={self._model_name}",
                f"--train_file={train_file_path}",
                "--energy_key=energy",
                "--forces_key=forces",
                f"--r_max={self._config.r_cut}",
                f"--max_num_epochs={self._config.num_epochs}",
                f"--lr={self._config.learning_rate}",
                "--device=cpu", # Forcing CPU for broader compatibility
                "--save_cpu", # Ensure model is saved in a CPU-compatible format
                "--E0s=average", # Provide required atomic energy references
                "--batch_size=1", # Use a batch size compatible with small test data
                "--valid_batch_size=1",
            ]

            # Execute the training command
            result = subprocess.run(command, capture_output=True, text=True)

            if result.returncode != 0:
                error_message = (
                    f"MACE training failed with exit code {result.returncode}.\\n"
                    f"Stdout:\\n{result.stdout}\\n"
                    f"Stderr:\\n{result.stderr}"
                )
                raise RuntimeError(error_message)

        finally:
            # Ensure the temporary file is always cleaned up
            if train_file_path.exists():
                train_file_path.unlink()

        # The actual model path is determined by MACE, using the 'name'
        final_model_path = self._work_dir / f"{self._model_name}.model"
        # The test expects a .pt file, so we will rename it for consistency
        # In a real scenario, we would just use the .model file.
        final_pt_path = self._work_dir / "models" / "trained_model.pt"
        final_pt_path.parent.mkdir(exist_ok=True)
        if final_model_path.exists():
            final_model_path.rename(final_pt_path)
            return final_pt_path

        raise FileNotFoundError("MACE did not produce the expected model file.")


    def _load_and_prepare_data(self, ids: list[int]) -> list[Atoms]:
        prepared_atoms_list = []
        for db_id in ids:
            row = self._db._db.get(id=db_id)
            dft_energy = row.energy
            dft_forces = row.forces
            atoms = self._db._db.get_atoms(id=db_id)

            target_energy = dft_energy
            target_forces = dft_forces

            if self._config.delta_learn:
                if self._config.baseline_potential == "lj":
                    baseline_energy = baseline_potentials.get_lj_potential(atoms)
                    baseline_forces = baseline_potentials.get_lj_forces(atoms)
                else:
                    raise ValueError("Unsupported baseline potential specified.")

                target_energy -= baseline_energy
                target_forces -= baseline_forces

            training_atoms = atoms.copy()
            # ASE's XYZ writer looks for results in the .info dictionary
            training_atoms.info['energy'] = target_energy
            training_atoms.arrays['forces'] = np.array(target_forces)

            prepared_atoms_list.append(training_atoms)

        return prepared_atoms_list
