pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomous-dev-env"
version = "0.1.0"
description = "AI-Native Cycle-Based Contract-Driven Development Environment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "python-dotenv",
    "pydantic-settings",
    "pydantic-ai>=0.0.18",
    "logfire>=2.0.0",
    "devtools",
    "langgraph>=1.0.5",
    "e2b-code-interpreter>=2.4.1",
    "textual>=6.10.0",
    "lancedb>=0.25.3",
    "sentence-transformers>=5.2.0",
    "tree-sitter==0.21.3",
    "tree-sitter-languages==1.10.2",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "google-auth>=2.43.0",
    "ase",
    "pyyaml",
    "mace-torch",
    "click",
]

[tool.uv]
environments = ["sys_platform == 'linux'"]

[project.scripts]
ac-cdd = "ac_cdd_core.cli:app"

[dependency-groups]
dev = [
    "typer",
    "rich",
    "shellingham",
    "pytest",
    "pytest-cov",
    "hypothesis",
    "playwright",
    "vcrpy",
    "ruff",
    "google-genai",
    "mypy",
    "bandit",
    "types-PyYAML",
    "pytest-asyncio>=1.3.0",
    "pre-commit>=4.5.1",
]

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.lint]
# E/F/I: Basic, B: Bugbear, S: Bandit, UP: pyupgrade
select = ["E", "F", "I", "B", "S", "UP"]
# S101: Allow use of assert (for tests)
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.pytest.ini_options]
addopts = "--cov=dev_src --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["dev_src/ac_cdd_core", "src/mlip_autopipec"]

[tool.mypy]
strict = true

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
targets = ["dev_src"]
